update riverpod last versions
